backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=backend/requirements.txt backend/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachecontrol==0.14.3
    # via firebase-admin
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
coverage[toml]==7.10.1
    # via pytest-cov
cryptography==45.0.5
    # via pyjwt
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
fastapi[all]==0.116.1
    # via -r backend/requirements.in
fastapi-cli[standard]==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.4
    # via fastapi-cli
firebase-admin==7.0.0
    # via -r backend/requirements.in
google-api-core[grpc]==2.25.1
    # via
    #   firebase-admin
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.21.0
    # via firebase-admin
google-cloud-storage==3.2.0
    # via firebase-admin
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
gunicorn==23.0.0
    # via -r backend/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx[http2]==0.28.1
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   firebase-admin
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.1
    # via cachecontrol
orjson==3.11.0
    # via fastapi
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic[email]==2.11.7
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.5
    # via fastapi
pydantic-settings==2.10.1
    # via fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt[crypto]==2.10.1
    # via firebase-admin
pytest==8.4.1
    # via
    #   -r backend/requirements.in
    #   pytest-cov
pytest-cov==6.2.1
    # via -r backend/requirements.in
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   -r backend/requirements.in
    #   fastapi
    #   uvicorn
requests==2.32.4
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.8
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
sentry-sdk==2.33.2
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typer==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.10.0
    # via fastapi
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
